update Items
create EquipmentSlotDao

+++ CS5200Project/src/main/java/game/model/Items.java
@@ -9,9 +9,9 @@ public class Items {
 	private String itemName;
 	private int level;
 	private int maxStackSize;
-	private BigDecimal price;
+	private double price;
 	
-	public Items(int itemID, String itemName, int level, int maxStackSize, BigDecimal price) {
+	public Items(int itemID, String itemName, int level, int maxStackSize, double price) {
 		super();
 		this.itemID = itemID;
 		this.itemName = itemName;
@@ -52,11 +52,11 @@ public class Items {
 		this.maxStackSize = maxStackSize;
 	}
 
-	public BigDecimal getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
